requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.0
channels==4.1.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
Django==5.1.1
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
pillow==10.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.2.1
service-identity==24.1.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==7.0.3
